Adjust package for auth
