+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/auth/user/Grantee.java
@@ -15,34 +15,21 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.auth;
+package org.apache.shardingsphere.infra.auth.user;
 
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
-import java.util.Collection;
-
 /**
- * ShardingSphere user.
+ * Grantee.
  */
 @RequiredArgsConstructor
 @Getter
-public final class ShardingSphereUser {
+@EqualsAndHashCode
+public final class Grantee {
     
     private final String username;
     
-    private final String password;
-    
     private final String hostname;
-    
-    private final Collection<String> authorizedSchemas;
-    
-    /**
-     * Get grantee.
-     *
-     * @return grantee
-     */
-    public Grantee getGrantee() {
-        return new Grantee(username, hostname);
-    }
 }
